backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
langchain==0.0.314
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
openai==1.3.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1 